requirements.txt
# Core Dependencies
opencv-python>=4.8.0
scipy>=1.10.0
numpy>=1.24.0
pandas>=2.0.0

# Optional: Better Face Detection
mediapipe>=0.9.0

# For Streamlit UI
streamlit>=1.34.0
streamlit-webrtc>=0.47.0
streamlit-mic-recorder>=0.0.4
gTTS>=2.5.0
matplotlib>=3.7.0
av>=10.0.0

# For FastAPI Backend
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
aiofiles>=23.0.0
python-multipart>=0.0.6

# Development/Testing (optional)
pytest>=7.4.0
requests>=2.31.0

# For Production Deployment
gunicorn>=21.0.0

# For AI Health Insights
groq>=0.4.0

# For PDF Reports
reportlab>=4.0.0
pillow>=10.0.0

# For User Authentication
bcrypt>=4.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
boto3>=1.28.0

